feat: Add Animal, AnimalHome, Dog, and Rabbit classes
This commit adds the Animal, AnimalHome, Dog, and Rabbit classes to the week3.sp1 package. The Animal class is an abstract class representing an animal with a name and the ability to make a sound. The Dog and Rabbit classes are subclasses of Animal, each with their own implementation of the makeSound method. The AnimalHome class is a new class that allows adding animals and retrieving a list of all animals. These changes introduce new functionality to the project.